v1.2
I was able to get the word guesser game to function as a second page, all you have to do is run node server.js hit log in and then click game 1, this still doesn't record the game score in the db or individual user sessions. I tried to start working on sessions but it looked like a whole lot of confusing stuff that i really don't understand i tried to use some code i found but it just completely broke everything when i tried it. with this it is almost as simple as plugging in the rest of the games to get them to function on the serve but untill we have sessions we cannot associate scores to a user specific user as far as i know.