GameCenterv1.3
Hey guys I changed a lot here, I added the snake game into the homepage, I think the leaderboard should work with mongoose for the snake game but I haven't updated the word game to output into the db yet. i also added a sign in (register) and login popup when you click the login button this should also work with the db but I didn't really test it out either if anybody wants to test this out and make any necessary fixes that would be cool. one thing that pissed me off a bit was when you open the log in/register popup the leaderboard clips into it and blocks some of the popup, I didnt take any time to investigate that but if someone can fix that it would be cool.